services/qr-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
boto3>=1.34.0
qrcode[pil]>=7.4.0
Pillow>=10.2.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.26.0
